# perf(async): Only allocate one oneshot per request