@badboy 2019-03-13
> (I wouldn't put to much stock in these)

To much stock into what?