@Marwes 2019-03-13
> What's the thrpt line in the benchmark output?

It is for `Throughput` (requests/second).

> To much stock into what?

Reworded it as `I wouldn't put to much stock in the exact numbers as they have pretty high variance, however this PR should always do less work than before`